Add files via upload
http://www.cburch.com/logisim/ ------ must download this free program to use.  
This is a 16-bit processor that has similar processor architecture to the 8080/z80 processor but with 16-bit registers. Virtual Processor includes: 16-bit registers (A,B,C,D,E,H,L), 64k of RAM, a keyboard, a TTY terminal, 32-bit by 32-bit LED matrix (with Address's mapped with register C, and data mapped with the D,E registers), and a mostly all the instructions from the 8080 instruction set. 

i also included a python script that will take a text file with 8080 instructions on it and assemble it to the machine code for the processor